More registration back out to specific client implementations
